Refactor: Implement dependency injection for dominance analysis
